fix bindgen issue on msvc
